Renamed privileged-service back to server
