+++ src/app/controller/new.ts
@@ -1,6 +1,6 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
-import { PrincipalService } from '../../principal/privileged-service';
+import { PrincipalService } from '../../principal/service';
 import { createAppForm } from '../formats/html';
 import { uuidUrn } from '../../crypto';
