+++ src/oauth2-client/controller/collection.ts
@@ -4,7 +4,7 @@ import { Context } from '@curveball/core';
 import { Forbidden, UnprocessableEntity } from '@curveball/http-errors';
 
 import * as hal from '../formats/hal';
-import { PrincipalService } from '../../principal/privileged-service';
+import { PrincipalService } from '../../principal/service';
 import { GrantType } from '../../types';
 import { OAuth2Client } from '../types';
 import { findByApp, create } from '../service';
