+++ src/one-time-token/controller/generate.ts
@@ -1,6 +1,6 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
-import { PrincipalService } from '../../principal/privileged-service';
+import { PrincipalService } from '../../principal/service';
 import { createToken } from '../service';
 import * as hal from '../formats/hal';
 import { resolve } from 'url';
