+++ src/user/controller/privileges.ts
@@ -4,7 +4,7 @@ import { BadRequest } from '@curveball/http-errors';
 import * as privilegeService from '../../privilege/service';
 import { PrivilegeMap } from '../../privilege/types';
 import * as hal from '../formats/hal';
-import { PrincipalService } from '../../principal/privileged-service';
+import { PrincipalService } from '../../principal/service';
 // import * as groupService from '../../group/service';
 
 type PolicyForm = {
